+++ lib/src/constants.dart
@@ -6,7 +6,7 @@ class Constants {
   static const String templateRepoUrl = 'https://github.com/nylo-core/nylo';
 
   /// Installer version
-  static const String version = '1.3.1';
+  static const String version = '1.4.0';
 
   /// Documentation URL
   static const String docsUrl = 'https://nylo.dev/docs';